backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
supabase>=2.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
google-adk>=0.1.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
mem0ai>=1.0.4
apscheduler>=3.10.0
nemoguardrails>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
redis>=5.0.0
fastembed>=0.2.0
fpdf2>=2.7.0
email-validator>=2.1.1
